core/validators.py
import magic
import re

from django.core.exceptions import ValidationError
from django.template.defaultfilters import filesizeformat
from django.utils.deconstruct import deconstructible


@deconstructible
class FileValidator(object):
    error_messages = {
     'max_size': ("Ensure this file size is not greater than %(max_size)s."
                  " Your file size is %(size)s."),
     'content_type': "Files of type %(content_type)s are not supported.",
    }

    def __init__(self, max_size=None, content_types=()):
        self.max_size = max_size
        self.content_types = content_types

    def __call__(self, data):
        if self.max_size is not None and data.size > self.max_size:
            params = {
                'max_size': filesizeformat(self.max_size),
                'size': filesizeformat(data.size),
            }
            raise ValidationError(self.error_messages['max_size'],
                                   'max_size', params)

        if self.content_types:
            content_type = magic.from_buffer(data.read(), mime=True)
            data.seek(0)

            if content_type not in self.content_types:
                params = { 'content_type': content_type }
                raise ValidationError(self.error_messages['content_type'],
                                   'content_type', params)

    def __eq__(self, other):
        return (
            isinstance(other, FileValidator) and
            self.max_size == other.max_size and
            self.content_types == other.content_types
        )

def validate_page_filter(value):

    assert value is None or isinstance(value, list)

    possible_value_re = re.compile("^\d+(-\d+)?$")
    valid_values = True

    for v in value:
        v = v.strip()
        if not re.match(possible_value_re, v):
            valid_values = False
            break

    if not valid_values:
        raise ValidationError(
            "Values must be a range or separate page numbers. E.g. 5, 10-14, 35", code="invalid"
        )
    else:
        return value
