requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==2.4.0
django-registration-redux==2.10
django-settings-export==1.2.1
django-simple-history==3.1.1
django-tables2==2.4.1
environs[django]==9.5.0
pillow==9.1.1
psycopg2-binary==2.9.3
python-magic==0.4.27
sqlparse==0.4.2
